Refactor all-time to items-learned and make more flexible
