+++ app/components/stats/items-learned/style.scss
@@ -1,6 +1,6 @@
 @import "../../../styles/_variables";
 
-gelato-component[data-name='stats-all-time'] {
+gelato-component[data-name='stats-items-learned-circle'] {
   .chart-overlay-wrapper {
     position: absolute;
     top: 25px;
