+++ app/components/stats/items-learned/template.jade
@@ -1,15 +1,16 @@
-gelato-component(data-name='stats-all-time')
+gelato-component(data-name='stats-items-learned-circle')
 
-  .title All Time
+  .title #{view.title}
   .row
     .col-md-3.center
       .characters.data-section
         #characters-learned.num-hero.chars-learned ...
         .units-label characters
         .colored-square.chars-learned-bg
-      .total-reviews.data-section
-        #total-reviews-num.num-subhero ...
-        .units-label total reviews
+      if view.showNumReviews
+        .total-reviews.data-section
+          #total-reviews-num.num-subhero ...
+          .units-label total reviews
     .col-md-6
       .circle
         #items-learned
@@ -22,8 +23,9 @@ gelato-component(data-name='stats-all-time')
         #words-learned.num-hero.words-learned ...
         .units-label words
         .colored-square.words-learned-bg
-      .days-studied.data-section
-        #total-time-studied-num.num-subhero ...
-        .units-label
-          span#units-total-label
-          span &nbsp; studied
+      if view.showDaysStudied
+        .days-studied.data-section
+          #total-time-studied-num.num-subhero ...
+          .units-label
+            span#units-total-label
+            span &nbsp; studied
