+++ app/components/stats/timeline/template.jade
@@ -24,7 +24,7 @@ gelato-component(data-name='stats-timeline')
       .items-leaned-days-studied-wrapper
         .row
           .col-md-6.middle-graph-wrapper-l
-            .items-learned-placeholder items learned
+            #items-learned-container
           .col-md-6.middle-graph-wrapper-r
             .days-studied-placeholder days studied
       .rwdt-graphs-wrapper
